Refactor to use contexts - still need to clean up but this is first pass. bump version 0.2.7
