+++ src/components/ReceiverConfig/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./ReceiverConfig";
+export { ReceiverContext } from "./ReceiverConfig";
